requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
dataclasses-json==0.6.1
Deprecated==1.2.14
deprecation==2.1.0
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
gotrue==1.2.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.310
langsmith==0.0.43
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.4
openai==0.28.1
orjson==3.9.7
packaging==23.2
pinecone-client==2.2.4
postgrest==0.11.0
posthog==3.0.2
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
realtime==1.0.0
regex==2023.10.3
requests==2.31.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-extensions==4.8.0
typing-inspect==0.9.0
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==10.4
wrapt==1.15.0
yarl==1.9.2
pytest
